review: content of input files inlined in test classes
