+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/BaseDeploymentApiHttpServiceBusinessSpec.scala
@@ -23,7 +23,6 @@ import pl.touk.nussknacker.test.config.{
 import pl.touk.nussknacker.ui.process.newdeployment.DeploymentId
 
 import java.io.File
-import java.nio.charset.StandardCharsets
 import java.nio.file.attribute.PosixFilePermissions
 import java.nio.file.{Files, Path}
 
@@ -67,34 +66,16 @@ trait BaseDeploymentApiHttpServiceBusinessSpec extends WithFlinkContainersDeploy
     )
 
   private lazy val inputDirectory = {
-    val rootDirectory = Files.createTempDirectory(s"nusssknacker-${getClass.getSimpleName}-transactions-")
-    Files.setPosixFilePermissions(rootDirectory, PosixFilePermissions.fromString("rwxr-xr-x"))
-    populateInputTransactionsDirectory(rootDirectory)
-    rootDirectory
+    val directory = Files.createTempDirectory(
+      s"nusssknacker-${getClass.getSimpleName}-transactions-",
+      // be default temp directory is read only for user
+      PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString("rwxr-xr-x"))
+    )
+    populateInputTransactionsDirectory(directory)
+    directory
   }
 
-  protected def populateInputTransactionsDirectory(rootDirectory: Path): Unit = {
-    val firstPartition = rootDirectory.resolve("date=2024-01-01")
-    firstPartition.toFile.mkdir()
-    FileUtils.write(
-      firstPartition.resolve("transaction-1.csv").toFile,
-      """"2024-01-01 10:00:00",client1,1
-        |"2024-01-01 10:01:00",client2,2
-        |"2024-01-01 10:02:00",client1,3
-        |""".stripMargin,
-      StandardCharsets.UTF_8
-    )
-    val secondPartition = rootDirectory.resolve("date=2024-01-02")
-    secondPartition.toFile.mkdir()
-    FileUtils.write(
-      secondPartition.resolve("transaction-1.csv").toFile,
-      """"2024-01-02 10:00:00",client1,1
-        |"2024-01-02 10:01:00",client2,2
-        |"2024-01-02 10:02:00",client1,3
-        |""".stripMargin,
-      StandardCharsets.UTF_8
-    )
-  }
+  protected def populateInputTransactionsDirectory(rootDirectory: Path): Unit
 
   private lazy val outputDirectory =
     Files.createTempDirectory(s"nusssknacker-${getClass.getSimpleName}-transactions_summary-")
