+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/DeploymentApiHttpServiceDeploymentCommentSpec.scala
@@ -15,6 +15,7 @@ import pl.touk.nussknacker.test.{NuRestAssureMatchers, RestAssuredVerboseLogging
 import pl.touk.nussknacker.ui.process.newdeployment.DeploymentId
 
 import java.nio.charset.StandardCharsets
+import java.nio.file.Path
 import scala.jdk.CollectionConverters._
 
 class DeploymentApiHttpServiceDeploymentCommentSpec
@@ -36,6 +37,29 @@ class DeploymentApiHttpServiceDeploymentCommentSpec
   override def designerConfig: Config = {
     super.designerConfig
       .withValue("deploymentCommentSettings.validationPattern", ConfigValueFactory.fromAnyRef(s".*$configuredPhrase.*"))
+  }
+
+  override protected def populateInputTransactionsDirectory(rootDirectory: Path): Unit = {
+    val firstPartition = rootDirectory.resolve("date=2024-01-01")
+    firstPartition.toFile.mkdir()
+    FileUtils.write(
+      firstPartition.resolve("transaction-1.csv").toFile,
+      """"2024-01-01 10:00:00",client1,1
+        |"2024-01-01 10:01:00",client2,2
+        |"2024-01-01 10:02:00",client1,3
+        |""".stripMargin,
+      StandardCharsets.UTF_8
+    )
+    val secondPartition = rootDirectory.resolve("date=2024-01-02")
+    secondPartition.toFile.mkdir()
+    FileUtils.write(
+      secondPartition.resolve("transaction-1.csv").toFile,
+      """"2024-01-02 10:00:00",client1,1
+        |"2024-01-02 10:01:00",client2,2
+        |"2024-01-02 10:02:00",client1,3
+        |""".stripMargin,
+      StandardCharsets.UTF_8
+    )
   }
 
   "The deployment requesting endpoint" - {
@@ -105,7 +129,6 @@ class DeploymentApiHttpServiceDeploymentCommentSpec
             }
             .verifyExternalState {
               val resultFile = getLoneFileFromLoneOutputTransactionsSummaryPartitionWithGivenName("date=2024-01-01")
-              // See test/resources/transactions to figure out why these values are expected
               FileUtils.readLines(resultFile, StandardCharsets.UTF_8).asScala.toSet shouldBe Set(
                 "client1,4",
                 "client2,2"
